Transferred Global Var and Functions to new JS File
Transferred global variable and the functions dependent on it from server to a new JS file (testcase-output.js)
